Update to Middy 3.0.1
This updates our Middy dependency to 3.0.1. Primarily, this requires
some refactoring of our `onError` handler, cleaning up the
`xss-santizer` tests, and removing an unneeded middleware.

A few other dependencies are updated as part of this and our logging
framework is added to a few more places (and we make use of its
middleware).