Update to Middy 3.0.1 (#862)
* Update to Middy 3.0.1

This updates our Middy dependency to 3.0.1. Primarily, this requires
some refactoring of our `onError` handler, cleaning up the
`xss-santizer` tests, and removing an unneeded middleware.

A few other dependencies are updated as part of this and our logging
framework is added to a few more places (and we make use of its
middleware).

* Fix dependency issue with IdP Constructs

We had some issues where in some scenarios, the IdP clients would get
created before the scopes and resource servers were done creating. This
does the most elegant thing it can but unfortunately it requires
exposing an extra bit of information from the Cognito IDP Client class.
But at least this didn't require any new functions or anything.

* Add a bunch of logging

* Consistently log across all functions

* Make the test bodies actually be JSON

* Cleanup import order