Add a TODO for better Access Key Secret structure
Everything inside of me wants to cry about a PR that just adds a TODO
but this seems like the best way to document "this bit is doing the best
thing that we can do today but not the best thing that might ever be
possible." It's not an urgent fix that we need either. But just
something that this feels like the best way to give context around this
piece of code.