Remove linting from pre-commit
For some team members, running `npm run lint` is remarkably slow (on the
order of minutes) when run across the full codebase. Because most
developers have ESLint run as part of the save process in their IDEs and
because we catch problems with linting in the CI pipeline; we can safely
remove it from the pre-commit configuration and be sure that unlinted
code does not make it to the `develop` branch. Once we've resolved
issues with linting being slow, we can potentially add this back.