Implement generate-document lambda (WIP)
Implement lambda to generate documents (e.g., DOW) that returns
a base64 encoded string of the pdf binary. Template files are
contained in a layer and are under the /templates folder. Juice
is a third-party package to combine the html and css files.

Still troubleshooting trying to get puppeteer to run inside of
lambda using puppeteer-core and having a compressed chromium
binary added as a dependency or some alternative. The
chrome-aws-lambda package does not appear to be maintained
anymore, but someone has forked the repo and has been updating
recent versions of node (v16) and puppeteer (v14).
- https://acloudguru.com/blog/engineering/serverless-browser-automation-with-aws-lambda-and-puppeteer
- https://github.com/alixaxel/chrome-aws-lambda
- https://github.com/Sparticuz/chrome-aws-lambda
- https://github.com/alixaxel/chrome-aws-lambda/pull/264
- https://github.com/alixaxel/chrome-aws-lambda/issues/275

Tikcet: AT-7345