Add bundling dep @sparticuz/chrome-aws-lambda to impl'd lambda
Use @sparticuz/chrome-aws-lambda as the chromium compressed binary
package since chrome-aws-lambda appears to no longer be maintained
but is seen in the community as the package to use. This is
slightly concerning, but works for now.

The puppeteer docs even suggest using chrome-aws-lambda for running
headless chrome on AWS Lambda.
- https://github.com/puppeteer/puppeteer/blob/main/docs/troubleshooting.md#running-puppeteer-on-aws-lambda

Ticket: AT-7345