Use new clearState features of logger
The Lambda Power Tools library now supports clearing "persistent" state
between invocations of the function. This is helpful because it means
that we can safely log things like user identity without a concern that
it would fail to be reset properly for the next invocation. This will
allow for logging.