Refactor IP Logging Middleware
The IP Logging Middleware previously logged the client/requester IP
address directly to standard output via `console.log`. This was
previously helpful for ensuring we had the information and was the best
thing we could do at the time it was implemented. As we have now added a
logging library (and because that library supports setting and clearing
state), we can actually now do something a little cooler.

This middleware is modified to add additional persistent state
attributes to the logger to add user identity information to _every_ log
message created during the execution of the function. We then apply the
middleware immediately after injecting the context for the logger and
(if needed clearing state) so that even the initial log message that
just dumps the event contains this. Having this data added in a
structured way means that we will be able to more easily query for this
information in CloudWatch and other tools.