Remove calls to console.log/error in app code
These calls either result in unstructured messages in the log output (if
they're part of actual application code) and/or unnecessary noise in
test output in the general case (if they're in app code or test code).
`console.log` family calls in the _CDK_ code are left as-is since that
operates more like a command-line application.

This will ensure that all messages have a structured format when in
CloudWatch Logs and hopefully make our test output slightly easier to
read.