Refactor CDK constructs
This refactors a pretty significant chunk of constructs within the
application to just directly extend other components rather than build
compositions based on them. For a few things, we're basically just
creating pre-configured variants of another construct type (SQS Queue,
State Machine, etc) and we can handle that pretty trivially and remove
some need to make wonky `.foo` attribute accesses. This also makes them
easier to just pass around within the CDK app.

Notably, the API Gateway REST API is not refactored. That adds a few
methods and my worry there is that we actually use some names that may
eventually conflict with the underlying resource (not likely but
possible) and so I kinda tried to avoid that.

Because the `SfnLambdaInvokeTask` was only ever created in one place, we
can also just get rid of it entirely. It wasn't doing a whole lot for
us. I think it made sense in a time where its use case was uncertain and
where we had more constructs with that pattern; today, I think we can
clean it up a bit.

This **will** result in resource replacement in already-deployed
environments as constuct paths (and therefore CloudFormation resource
logical IDs) will change. I feel that because these are mostly Queues
and connecting resources this is safe but there's a chance this will gum
up the pipeline and we'll either need to make some additional fixes or
revert.