Log all DNS queries to CloudWatch Logs
This will ensure that we have a record of every DNS query made within
our VPC, regardless of which resource made the request (Lambda Function
or other resource). So long as it came from a ENI and hit the Amazon
default DNS server, we'll have it. These logs go to a CloudWatch Logs
Log group.

We have to use an L1 resource for this because the CDK does not provide
L2 resources.