Add required notifications to be sent via email
Similar to the currently-open monitoring PR, this adds a base for
setting up notifications everywhere except our sandbox account. These
notifications are required under the STIGs and help ensure we meet some
requirements under 800-53.

We ignore Sandbox because we build IAM stuff there all the day and we'll
never see a real email; it'd just be notifications.