Add an optional stack for network troubleshooting
This adds a stack that basically almost never gets built. It in order to
set this, `atat:NetworkTroubleshootingEnvironment` must be set to
"true", "1", or "yes". The created stack basically just creates an EC2
instance that allows port 80 and ICMP from anywhere. This is sufficient
for testing basic connectivity from a firewall or bastion and for
bootstrapping any tests for NAT.