Support custom domains for the API Gateway
This adds support for custom domains to the API Gateway. The way that
this is handled requires changes to a few core constructs within the
environment as well as a little bit of gluing AWS resources together.
First, API Gateway does not directly support custom domains on private
endpoints. This requires us to put _something_ in front of it that can
speak HTTPS with a valid certificate for a custom name. This uses a load
balancer. Because of issues experienced during troubleshooting, it was
easier to just add support for _both_ types of Load Balancers and we can
toggle between them in the future as needed.

Now, as for how we implement the Load Balancer's Target Group... that
requires a little more involvement. To start with, ELBv2 only supports a
few types of targets by default: EC2 Instances, IP Addresses, Lambda
Functions, and ALBs. Unfortunately, a VPC Endpoint (which is what we
have with our API Gateway Private Endpoints) isn't any of those things
directly. But it does have an IP address (well... it actually has
several). Unfortunately, no tooling directly exposes the IP addresses
based on VPC Endpoint ID; especially not CloudFormation/CDK.
Fortunately, with Custom Resources, we can glue together several AWS API
calls to describe the endpoint and get its associated Elastic Network
Interfaces (ENIs) and then describe those to get their IP addresses
(because despite ENIs being a lovely abstraction in EC2/AWS, an ELBv2
cannot target them).

Unfortunately, the CDK is very helpful. And by being so helpful, it
provides a lot of abstractions around ELBv2 Target Groups. But this
breaks the ability to just yeet arbitrary JSON into the `Targets`
property of the `AWS::ElasticLoadBalancingV2::TargetGroup` resource. No
matter what you try to do, the CDK either wraps it in a list or very
(not-so) helpfully tries to validate it (which won't work because with
our custom resource, it ends up being a `Fn::GetAtt`). And because it's
an intrinsic function and not actual raw data, we can't treat it like an
object and split it, iterate over it, or any other fun magic (in the
CDK, it's a `Token`). So we have to just treat it as an opaque object.
The only way we can effectively provide that as an opaque object is to
use escape hatches in the CDK to pass it directly to the underlying
CloudFormation resource property. That means that our Custom Resource
has to return a valid JSON object that can be passed to the `Targets`
object.

The configured Load Balancer will then balance across all the IP address
(and hopefully does so somewhat intelligently based on availability zone
since we do provide that data). The HTTPS/TLS certificate from ACM is
attached to the load balancer and that results in the certificate
matching the DNS name.

This does require two changes for consumers of the API:

 1. Consumers must now use the DNS name. The regional endpoint for the
    API is no longer available
 2. Consumers must pass the `x-apigw-api-id` header, as described in the
    [API Gateway docs for invoking via the DNS name][1]

Finally, this should not cause any issues for Developers' sandbox
environments. Those should be able to trivially be built without DNS
names and without VPC connectivity.

This implementation does have a few trade-offs as it closely ties the
usage of a custom domain to the usage of a VPC. For our use case, that
is likely pretty acceptable since any environments where we'd specify a
custom name, we'd also have a VPC. Second, we may need to kick off
something to re-trigger the custom resource at some point in the future
if we for any reason see the underlying IPs change without the VPCE ID
changing (but that is extremely unlikely to happen).

One option that was available that was rejected for now was having the
ALB target a Lambda to handle balancing between the VPC endpoints. This
would have resulted in additional code needing to be executed on _every_
request (so in many cases, doubling our number of Lambda function
invocations) and would have just added another link in the chain of
request/response size limitations that would become somewhat harder to
debug. Not using the Lambda option also allows us to maintain the
ability to use either an NLB or an ALB, whichever is most appropriate.

[1]: https://docs.aws.amazon.com/apigateway/latest/developerguide/apigateway-private-api-test-invoke-url.html#apigateway-private-api-public-dns