Resolve or supress all 800-53r4 checks in ATAT stack
We still need to do some work to solve IAM and Network; however, this
fixes the issues in the main app stack. All remaining issues are
resolved or are suppressed and the check is added to the deployment
pipeline to ensure that we always remain compliant. This is far easier
to do now that we have networking solved (except that tricky
`AwsCustomResource`).