Use the new `userPoolClientSecret` property
v2.37.0 of the AWS CDK added the ability to get a Client Secret directly
via an AWS Custom Resource. So there's not really a need for us to
create our own any more. We _were_ doing this to also get the client's
"ClientName" attribute but we never used it and _that_ certainly isn't
worth the maintenance burden of this whole construct.