Update igce-document fn impl
Update fn impl to add additional rows on the period sheets or
summary sheets as needed. This is done by duplicating rows and
updating formulas on the period sheets and summary sheets.

The summary sheet total and grand total calculations is added
manually due to an error by exceljs related to sharedType formulas.
The calculations are adjusted to account for the additional rows
on the summary sheet.

Ticket: AT-7759