Add ability to override GitHub PAT secret version
In some scenarios, the PAT needs to be rotated; however, CloudFormation
will not automatically pick up that there's a new version. This is
because CloudFormation will only re-resolve the secret when either the
version changes or a property of the resource changes. I tried to do
this with Tags but that didn't seem to be sufficient.

Instead, what we do is conditionally pass a specific version ID to
resolve when the `atat:ForceGitHubTokenVersion` context value is
provided. When not provided, this will be `undefined` and the default
value will be used (`AWSCURRENT`). Otherwise, the given value will be
used.

This means that we need to _manually_ run a single CDK deploy in each
environment after rotating the GitHub PAT. CloudFormation will update
the pipeline and the pipeline will then automatically update itself to
not specify a particular version. This parameter should be provided
manually on the CLI and not via `cdk.json` (nor via `cdk.context.json`)
to allow it to just be specified for the one execution.