Refactor context key/value definitions
This extracts all currently used context values into a shared enum-like
class that will allow for easier and more consistent resolution of
context values as well as, most excitingly, clear documentation for each
of the used context values.

Each of the given context values is extracted to an instance of the
`AtatContextValue` class and exposed as a static variable of the class;
the class has a private constructor and all attributes are readonly to
limit runtime modifications. The static properties are documented to
describe the intended behavior and usage of the context key and how the
value will be used. A helper is provider to allow automatically
resolving within a given scope in the CDK app as well the ability to
fallback to the default value (if one exists). A helper is also created
to support converting the values to CLI arguments (helpful in the
pipeline).

The test change is just a verbiage modification for Sandbox
environments; otherwise, this does not have any significant changes on
the output of the application, either via CLI or synthesized output.