Update AWS CDK to v2.41.0 to put AwsCustomResource in VPC
The new version of the CDK adds support for putting `AwsCustomResource`
resources into a VPC. Because of the underlying ways that these are
created, we just need to create one that ends up having the base
configuration shared across all AwsCustomResource constructs created by
us and elsewhere in the tree, so long as this one is created first
(which is why it's at the top of the `Stack`).

This _only_ makes the changes in the `atat-web-api` stack because that's
the only place where we know the VPC is ready (for example, we can't do
this in the VPC stack because the most important custom resource is
created before the VPC is).