Add support for asynchronous requests (#1185)
This adds support for handing HTTP 202 responses to provisioning
requests. This is accomplished by trying to cleanup a bit of the logic
and handling and to remove some assumptions.

Primarily, this is complicated by the fact that it is not always the
case that the HOTH and ATAT APIs use the same structure for objects. In
fact, this assumption has been made throughout the code base; and it's
incorrect. We cannot just pass up `payload` objects because the ATAT API
requires `snake_case` and all HOTH objects are `camelCase`. To help with
this, a standard ATAT client implementation is added in `api/client`
that handles transforming between the two. This is implemented with
`lodash` and `axios`; the former dependency feels a bit heavy for this
but it gets the job done.

Second, a lot of type safety is added here. There are a few places where
I had to fall back to using `any` in order to actually be able to get
something done. But it's a lot better than it used to be.

During this, it became clear that there's a lot of assumptions in the
case base that we should trust the `targetCsp` attribute and we
shouldn't. We should use the name, fetch the configuration, and get
everything else from there. All non-`name` fields in
`CloudServiceProvider` have been deprecated and made optional.

The handler that setup the step function execution used to do some stuff
to normalize the request for more easily passing it on to the CSPs; this
is no longer necessary with the client (and it was somewhat incomplete
previously). Most of the code is removed but the handler remains to give
us flexibility. In the future, we should investigate using a more direct
APIGw->SFN integration.

There are still some tests that need to be written here but (at least,
locally) all existing tests pass except for some that are presently
skipped for the provisioning workflow; however, it should be possible to
follow the same general process that is used for the cost functions.

Ticket: AT-7902
Ticket: AT-7903
Ticket: AT-7904

Co-authored-by: Julius Fitzhugh (Hunter Strategy) <julius.fitzhugh-ctr@ccpo.mil>