Update logger type to any to prevent tests from failing
There appears to be difference between the logger type for
sdk clients in @aws-sdk/type and @aws-lambda-powertools/logger
that causes tests using one of the sdk clients to fail.

This is a temp fix until the issue is resolved and this can
be reverted. The logger appears to be working when testing
against a HOTH API endpoint (provisioning-jobs).

See files:
- node_modules/@aws-sdk/types/dist-types/logger.d.ts
- node_modules/@aws-lambda-powertools/logger/lib/Logger.d.ts