Add templates to support eval plan...
Adds 7 templates to support eval plan.
eval plan isn't a simple template, each page has unique
hardcoded bullet point lists and full Evaluation Tables, it could be possible to generate
this all in one template but the formatting got very messy, you can't just keep the table in the
template with an if, you have to actually format it with javascript built in, and it wasn't rendering.

Theoretically, you could use HTML to generate these tables and full basis of evaluation lists, due
to time constraint this achieves the same functionality.

Ticket:AT-8374