Add ELB Access Logs
These logs will be helpful from a troubleshooting perspective as well as
required for compliance and security reasons. Enabling them requires
specifying the region on the stack. This is really unfortunate because
it means that our template isn't 100% portable anymore (as in, an
already-built template can no longer be run in any region; though, we
can still build a template that works in any region). The region can be
overriden by specifying the `atat:DeployRegion` value. The default value
is `"us-gov-west-1"` but can be changed to allow deploying to another
region.

Additionally, minor changes are made to clean up additional synthesis
errors and warnings that have started to appear (either as a result of
this change or that just happened to be noticed now because of looking
more closely for issues).