Update dow docx template
Add 'Statement of objectives for anticipated need/usage, to include
purpose and usage of the expected tools/capabilities:' to the
following XaaS services based on DITCO request.
- Developer Tools
- Applications
- Machine Learning
- Networking
- Security
- Edge Compute
- IoT

Add a comment at the top of the document to add helpful tips
regarding the template.

Remove unused assignments previously used for section 5 table.