feature/AT-9429-axios-interceptor
Updated interceptor to check header content type for 'starts with' rather than strict equality
so as to handle 'application/json' as well as 'application/json' with a specified charset.

Added two unit tests; handle application/json with utf-8 as well as a test to ensure
it doesn't throw when not given data.